update structure file (#8)
